pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyiota"
dynamic = ["version"]
description = "Python SDK for IOTA Rebased"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Consensix Labs", email = "contact@consensixlabs.com" },
]
keywords = ["iota", "blockchain", "move", "sdk", "web3", "rebased"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "pynacl>=1.5",
]

[project.optional-dependencies]
secp256k1 = ["coincurve>=20.0"]
mnemonic = ["mnemonic>=0.21"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.34",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.urls]
Homepage = "https://github.com/Consensix-Labs/pyiota"
Documentation = "https://github.com/Consensix-Labs/pyiota#readme"
Repository = "https://github.com/Consensix-Labs/pyiota"
Issues = "https://github.com/Consensix-Labs/pyiota/issues"
Changelog = "https://github.com/Consensix-Labs/pyiota/blob/main/CHANGELOG.md"

[tool.hatch.version]
path = "src/pyiota/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/pyiota"]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "TCH", "RUF"]

[tool.ruff.lint.isort]
known-first-party = ["pyiota"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Unit tests (no network required)",
    "integration: Integration tests (requires testnet)",
]
filterwarnings = ["error"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
